Move GeocodingServiceProvider out of Silex namespace
